renamed form.php to better describe what page it was for
