+++ add_item.php
@@ -58,4 +58,4 @@ if ($_POST) {
 }
 
 // Display simple form for user
-require 'form.php';
+require 'add_item_form.php';
